+++ examples/pirates/rules.rs
@@ -24,6 +24,7 @@ impl TeamRules<PiratesRules> for PiratesTeamRules {
     // Teams don't have powers in this example.
     type Power = EmptyPower;
     type PowersSeed = ();
+    type Invocation = ();
     type PowersAlteration = ();
     // We we'll use the id of the opposing team both seed and objective, to check if the
     // goal of sinking the enemy ship was achieved.